Move directory path value objects and tests under Path namespace to reflect responsibility
