+++ src/Path/ValidatedDirectoryPath.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace Haspadar\Piqule\File;
+namespace Haspadar\Piqule\Path;
 
-use Haspadar\Piqule\Path\Path;
 use Haspadar\Piqule\PiquleException;
+use Override;
 
 final readonly class ValidatedDirectoryPath implements Path
 {
@@ -13,6 +13,7 @@ final readonly class ValidatedDirectoryPath implements Path
         private DirectoryPath $origin,
     ) {}
 
+    #[Override]
     public function value(): string
     {
         $value = $this->origin->value();
