+++ tests/Unit/Path/DirectoryPathTest.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Haspadar\Piqule\Tests\Unit\File;
+namespace Haspadar\Piqule\Tests\Unit\Path;
 
-use Haspadar\Piqule\File\DirectoryPath;
+use Haspadar\Piqule\Path\DirectoryPath;
 use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;
