+++ tests/Unit/Path/ValidatedDirectoryPathTest.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace Haspadar\Piqule\Tests\Unit\File;
+namespace Haspadar\Piqule\Tests\Unit\Path;
 
-use Haspadar\Piqule\File\DirectoryPath;
-use Haspadar\Piqule\File\ValidatedDirectoryPath;
+use Haspadar\Piqule\Path\DirectoryPath;
+use Haspadar\Piqule\Path\ValidatedDirectoryPath;
 use Haspadar\Piqule\PiquleException;
 use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;
